Rename SchemaParser to JsonSchemaParser because it is the only component specific to JSON schemas
